World Map and Map Generation
You can now walk through the world now! The map is bigger and there's a function in the Centre script which allows you to generate maps of your own! All you have to do is insert an mage, change the directory and copy and paste the output numbers into a textfile