Map generator tweak
I tried to make it so that editing levels can be done easier without having to copy and paste text. All you gotta' do now is just change the directory and edit an image file and then you should be good to go. Only problem is that it doesn't update immediately so you got to load the game, close it and reload it again for it to work. not too much of a hassle I'll say but might limit progress a little