Big Changes
Hehe.. I forgot to upload! Anyways I did a little bit of customization off camera. I got a few more items and implemented UI. I dont know if I did that yet but its certainly here now. The screen is bigger too and the UI scales to it. Player is customizeable as can be seen with the blue stick figure. Little bit of work but I can always fix it.