Mostly adjusting sprites
Items now have a blue border to distinguish them from the rest of the map. I originally didn't want to do this because I wanted everything to be more realistic and stuff like that but it now makes items easier to see. Everything is small though so I might change that later